Refactor view, viewmodel and model
