+++ src/Xalendar.Sample/Xalendar.Sample/Models/CustomEvent.cs
@@ -1,0 +1,23 @@
+﻿using System;
+using Xalendar.Api.Interfaces;
+
+namespace Xalendar.Sample.Models
+{
+    public class CustomEvent : ICalendarViewEvent
+    {
+        public object Id { get; }
+        public string Name { get; }
+        public DateTime StartDateTime { get; }
+        public DateTime EndDateTime { get; }
+        public bool IsAllDay { get; }
+        
+        public CustomEvent(object id, string name, DateTime startDateTime, DateTime endDateTime, bool isAllDay)
+        {
+            Id = id;
+            Name = name;
+            StartDateTime = startDateTime;
+            EndDateTime = endDateTime;
+            IsAllDay = isAllDay;
+        }
+    }
+}
